Big refactoring

+++ src/main/kotlin/com/n26/transactionstatistics/statistics/domain/Statistics.kt
@@ -2,4 +2,4 @@ package com.n26.transactionstatistics.statistics.domain
 
 import java.math.BigDecimal
 
-data class Statistics(val sum: BigDecimal, val avg: BigDecimal, val max: BigDecimal, val min: BigDecimal, val count: Int)
+data class Statistics(val sum: BigDecimal, val avg: BigDecimal, val max: BigDecimal, val min: BigDecimal, val count: Long)
