+++ src/main/kotlin/com/n26/transactionstatistics/statistics/infra/adapters/primary/dto/StatisticsAPI.kt
@@ -1,3 +1,3 @@
 package com.n26.transactionstatistics.statistics.infra.adapters.primary.dto
 
-data class StatisticsAPI(val sum: String, val avg: String, val max: String, val min: String, val count: Int)
+data class StatisticsAPI(val sum: String, val avg: String, val max: String, val min: String, val count: Long)
